# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MechMind OS API
  description: |
    MechMind OS v10 - Multi-tenant SaaS for automotive repair shops with AI voice booking.
    
    ## Features
    - AI-powered voice booking via Vapi integration
    - Multi-tenant architecture with row-level security
    - Real-time slot availability with advisory locks
    - Webhook-based voice call handling
    - GDPR-compliant data management
    
    ## Authentication
    All API endpoints require Bearer token authentication. Obtain tokens via the `/v1/auth/login` endpoint.
    
    ## Rate Limiting
    - Standard tier: 100 requests/minute
    - Premium tier: 500 requests/minute
    - Enterprise tier: 2000 requests/minute
    
    ## Base URL
    - Production: `https://api.mechmind.io`
    - Staging: `https://api-staging.mechmind.io`
    - Sandbox: `https://api-sandbox.mechmind.io`
  version: 1.0.0
  contact:
    name: MechMind Support
    email: api-support@mechmind.io
    url: https://docs.mechmind.io
  license:
    name: Proprietary
    url: https://mechmind.io/terms
servers:
  - url: https://api.mechmind.io/v1
    description: Production server
  - url: https://api-staging.mechmind.io/v1
    description: Staging server
  - url: https://api-sandbox.mechmind.io/v1
    description: Sandbox server
security:
  - BearerAuth: []
...
paths:
...
  /v1/bookings:
    get:
      summary: List bookings
      description: |
        Retrieve a paginated list of bookings. Supports filtering by date range,
        mechanic, customer, and status. Results are scoped to the authenticated
        tenant (shop).
      tags:
        - Bookings
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: start_date
          in: query
          schema:
            type: string
            format: date
          description: Filter bookings from this date (inclusive)
        - name: end_date
          in: query
          schema:
            type: string
            format: date
          description: Filter bookings until this date (inclusive)
        - name: mechanic_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by assigned mechanic
        - name: customer_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by customer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, in_progress, completed, cancelled, no_show]
          description: Filter by booking status
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    post:
      summary: Create booking
      description: |
        Create a new booking. The system will validate slot availability
        and assign to the specified mechanic. Requires `bookings:write` scope.
      tags:
        - Bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Slot unavailable or conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: SLOT_UNAVAILABLE
                  message: "The requested time slot is no longer available"
                  details:
                    slot_id: "550e8400-e29b-41d4-a716-446655440000"
                    suggested_alternatives:
                      - slot_id: "550e8400-e29b-41d4-a716-446655440001"
                        start_time: "2024-01-15T10:00:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /v1/bookings/reserve:
    post:
      summary: Reserve slot with advisory lock
      description: |
        Reserve a time slot using PostgreSQL advisory locks for concurrency control.
        This endpoint is used by the voice booking system to prevent double-booking
        during the conversation flow. Lock is held for 5 minutes.
      tags:
        - Bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - slot_id
                - mechanic_id
                - customer_phone
              properties:
                slot_id:
                  type: string
                  format: uuid
                  description: ID of the time slot to reserve
                  example: "550e8400-e29b-41d4-a716-446655440000"
                mechanic_id:
                  type: string
                  format: uuid
                  description: ID of the mechanic
                  example: "550e8400-e29b-41d4-a716-446655440001"
                customer_phone:
                  type: string
                  description: Customer phone number (E.164 format)
                  example: "+14155551234"
                reservation_duration:
                  type: integer
                  default: 300
                  description: Lock duration in seconds (max 600)
      responses:
        '201':
          description: Slot reserved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  reservation_id:
                    type: string
                    format: uuid
                  expires_at:
                    type: string
                    format: date-time
                  slot:
                    $ref: '#/components/schemas/TimeSlot'
        '409':
          description: Slot already reserved or booked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: SLOT_LOCKED
                  message: "Slot is currently reserved by another process"
                  details:
                    reserved_until: "2024-01-15T09:05:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /v1/bookings/{booking_id}:
    parameters:
      - name: booking_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: Unique booking identifier

    get:
      summary: Get booking details
      description: Retrieve detailed information about a specific booking
      tags:
        - Bookings
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      summary: Update booking
      description: |
        Update booking details. Cannot modify completed or cancelled bookings.
        Status changes trigger appropriate notifications.
      tags:
        - Bookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBookingRequest'
      responses:
        '200':
          description: Booking updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Cancel booking
      description: |
        Cancel a booking. Sends cancellation notification to customer.
        Completed bookings cannot be cancelled.
      tags:
        - Bookings
      parameters:
        - name: reason
          in: query
          schema:
            type: string
          description: Cancellation reason
      responses:
        '204':
          description: Booking cancelled successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: Cannot cancel completed booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/slots:
    get:
      summary: List available time slots
      description: |
        Get available time slots for booking. Returns slots filtered by
        date range, mechanic, and service type. Includes real-time availability.
      tags:
        - Slots
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Date to query (YYYY-MM-DD)
        - name: mechanic_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by specific mechanic
        - name: service_type
          in: query
          schema:
            type: string
          description: Filter by service type duration requirements
        - name: duration_minutes
          in: query
          schema:
            type: integer
            minimum: 15
            maximum: 480
          description: Required slot duration in minutes
      responses:
        '200':
          description: Available time slots
          content:
            application/json:
              schema:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                  slots:
                    type: array
                    items:
                      $ref: '#/components/schemas/TimeSlot'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /v1/auth/login
    
    VapiWebhookAuth:
      type: apiKey
      in: header
      name: X-Vapi-Signature
      description: HMAC-SHA256 signature of webhook payload
    
    MetricsAuth:
      type: http
      scheme: bearer
      description: Metrics API key for Prometheus scraping

  # ==================== PARAMETERS ====================
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number for pagination
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Items per page

  # ==================== RESPONSES ====================
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: "Valid authentication credentials required"
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: FORBIDDEN
              message: "Insufficient permissions for this operation"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: "Requested resource not found"
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
  schemas:
...
    CreateBookingRequest:
      type: object
      required:
        - slot_id
        - mechanic_id
        - customer_phone
      properties:
        slot_id:
          type: string
          format: uuid
        mechanic_id:
          type: string
          format: uuid
        customer_phone:
          type: string
        customer_name:
          type: string
        customer_email:
          type: string
          format: email
        vehicle_info:
          $ref: '#/components/schemas/VehicleInfo'
        service_type:
          type: string
        notes:
          type: string
        source:
          type: string
          enum: [web, voice, phone, walkin, api]
          default: api
    UpdateBookingRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled, no_show]
        mechanic_id:
          type: string
          format: uuid
        service_type:
          type: string
        notes:
          type: string
        vehicle_info:
          $ref: '#/components/schemas/VehicleInfo'
    BookingResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        shop_id:
          type: string
          format: uuid
        slot_id:
          type: string
          format: uuid
        mechanic_id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        customer:
          $ref: '#/components/schemas/CustomerBrief'
        mechanic:
          $ref: '#/components/schemas/MechanicBrief'
        slot:
          $ref: '#/components/schemas/TimeSlot'
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled, no_show]
        service_type:
          type: string
        vehicle_info:
          $ref: '#/components/schemas/VehicleInfo'
        notes:
          type: string
        source:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        confirmed_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    BookingListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BookingResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ----- Slot Schemas -----
    TimeSlot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        shop_id:
          type: string
          format: uuid
        mechanic_id:
          type: string
          format: uuid
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        is_available:
          type: boolean
        is_reserved:
          type: boolean
        reserved_until:
          type: string
          format: date-time
        booking_id:
          type: string
          format: uuid

    # ----- Mechanic Schemas -----
...
    MechanicBrief:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        specialties:
          type: array
          items:
            type: string
...
    CustomerBrief:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        email:
          type: string
...
    VehicleInfo:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        vin:
          type: string
        license_plate:
          type: string
        color:
          type: string

    # ----- Shop Schemas -----
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            error:
              type: object
              properties:
                validation_errors:
                  type: array
                  items:
                    type: object
                    properties:
                      field:
                        type: string
                      message:
                        type: string
                      code:
                        type: string